Move keypress listener into menu
