+++ src/helpers.js
@@ -3,3 +3,10 @@ export function getRandomInt(min = 0, max = 1) {
   max = Math.floor(max + 1);
   return Math.floor(Math.random() * (max - min)) + min;
 }
+
+export function playSound(soundKey, startPoint = 0, volume = 1) {
+  const audio = document.querySelector(`[data-sound=${soundKey}]`);
+  audio.currentTime = startPoint;
+  audio.volume = volume;
+  audio.play();
+}
